requirements.txt
streamlit==1.35.0
biopython==1.83
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
